requirements.txt
# Core dependencies
rich>=13.0.0
textual>=0.40.0
aiohttp>=3.8.0
aiosqlite>=0.19.0

# AI and vector storage
ollama>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Data handling
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Game mechanics
dice>=4.0.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0